Tweaked lighting, added grass texture for testing
The lighting on the grass (seaweed/kelp) is mostly right, but we have some issues. The normals and tangents are not correct. I suspect that is why the lighting is behaving strange for those plants. Its the cylinder itself that has the issue. Not really a priority to fix, it looks fine for the most-part